+++ test-utils/src/main/java/io/ballerina/stdlib/http/testutils/nativeimpl/Stderr.java
@@ -1,0 +1,43 @@
+/*
+ *  Copyright (c) 2026 WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
+ *
+ *  WSO2 LLC. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package io.ballerina.stdlib.http.testutils.nativeimpl;
+
+import io.ballerina.runtime.api.values.BObject;
+import io.ballerina.stdlib.io.channels.AbstractNativeChannel;
+import io.ballerina.stdlib.io.channels.BlobChannel;
+import io.ballerina.stdlib.io.channels.BlobIOChannel;
+
+import java.io.InputStream;
+import java.nio.channels.Channels;
+import java.nio.channels.ReadableByteChannel;
+
+/**
+ * External function for Process.stderr.
+ *
+ * @since 2.16.0
+ */
+public class Stderr extends AbstractNativeChannel {
+
+    public static BObject stderr(BObject objVal) {
+        Process process = OSUtils.processFromObject(objVal);
+        InputStream in = process.getErrorStream();
+        ReadableByteChannel readableByteChannel = Channels.newChannel(in);        
+        return createChannel(new BlobIOChannel(new BlobChannel(readableByteChannel)));
+    }
+}